Made single-header, renamed namespace to sacfmt, renamed SacStream to Trace
