+++ CMakeLists.txt
@@ -5,7 +5,7 @@
 cmake_minimum_required (VERSION 3.12)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 # Build type
-set(CMAKE_BUILD_TYPE Debug)
+set(CMAKE_BUILD_TYPE Release)
 if (NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE Release)
 endif()
@@ -39,8 +39,8 @@ set(CMAKE_CXX_FLAGS_RELEASE "-O3 -DNDEBUG")
 #==============================================================================
 project(sac-format
   LANGUAGES CXX
-  DESCRIPTION "Seismic Analaysis Code Binary File Library"
-  HOMEPAGE_URL https://github.com/arbCoding/sac-format
+  DESCRIPTION "C++20 Single-Header SAC File library"
+  HOMEPAGE_URL https://arbCoding.github.io/sac-format
   VERSION 0.9.0)
 # Use C++20
 set(CMAKE_CXX_STANDARD 20)
@@ -48,23 +48,28 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
 add_library(sac-format STATIC
-  src/sac_io.hpp
-  src/sac_io.cpp
-  src/sac_stream.hpp
-  src/sac_stream.cpp)
+  src/sac_format.hpp
+  src/sac_format.cpp)
 
 set_target_properties(sac-format PROPERTIES VERSION ${PROJECT_VERSION})
-set_target_properties(sac-format PROPERTIES PUBLIC_HEADER src/sac_stream.hpp)
+set_target_properties(sac-format PROPERTIES PUBLIC_HEADER src/sac_format.hpp)
 #===============================================================================
 # Unit tests
 #===============================================================================
 # Catch2 stuff
-# Doesn't like Weffc++
 add_executable(utests
-  src/sac_stream.hpp
+  src/sac_format.hpp
   src/utests.cpp)
 
 target_link_libraries(utests
+  PRIVATE Catch2::Catch2WithMain
+  sac-format)
+
+add_executable(benchmark
+  src/sac_format.hpp
+  src/benchmark.cpp)
+
+target_link_libraries(benchmark
   PRIVATE Catch2::Catch2WithMain
   sac-format)
